Docstrings

AcousticFeatures.Score — Method

Score(
    f::AcousticFeatures.AbstractAcousticFeature,
    x::MetaArrays.MetaArray{var"#s13",SignalAnalysis.SamplingInfo,T<:Real,N} where N where var"#s13";
    winlen,
    noverlap,
    preprocess,
    showprogress
) -> Any

Compute acoustic feature f scores of a SampledSignal x using sliding windows.

By default, window length winlen is the length of x, i.e., the whole signal is used to compute a score, and overlapping samples noverlap is 0. The padtype specifies the form of padding, and for more information, refer to ImageFiltering.jl. The signal is subject to preliminary processing preprocess. Acoustic feature scores of subseqences can be computed through mapping map. showprogress is used to monitor the computations.

AcousticFeatures.Score — Method

Score(
    f::AcousticFeatures.AbstractAcousticFeature,
    x::Union{AbstractArray{T<:Real,1}, AbstractArray{T<:Real,2}};
    fs,
    winlen,
    noverlap,
    preprocess,
    showprogress
) -> Any

Compute acoustic feature f scores of a signal x using sliding windows.

By default, sampling rate fs is one, window length winlen is the length of x, i.e., the whole signal is used to compute a score, and overlapping samples noverlap is 0. The padtype specifies the form of padding, and for more information, refer to ImageFiltering.jl. The signal is subject to preliminary processing preprocess. Acoustic feature scores of subseqences can be computed through mapping map. showprogress is used to monitor the computations.

AcousticFeatures.score — Method

score(
    f::Myriad{S<:Real},
    x::AbstractArray{T<:Real,1}
) -> Array{_A,1} where _A

Score of x based on myriad.

Reference:

Mahmood et. al., "Optimal and Near-Optimal Detection in Bursty Impulsive Noise," IEEE Journal of Oceanic Engineering, vol. 42, no. 3, pp. 639–653, 2016.

Examples:

julia> x = Score(Myriad(), randn(9600))
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1]
    :feature, ["Myriad"]
    :channel, [1]
And data, a 1×1×1 Array{Float64, 3}:
[:, :, 1] =
 30884.887026311182

julia> x = Score(Myriad(), randn(9600); winlen=960, noverlap=480)
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1, 481, 961, 1441, 1921, 2401, 2881, 3361, 3841, 4321, 4801, 5281, 5761, 6241, 6721, 7201, 7681, 8161, 8641]
    :feature, ["Myriad"]
    :channel, [1]
And data, a 19×1×1 Array{Float64, 3}:
[:, :, 1] =
 34705.27101918274
 ⋮
 34616.48633137812

AcousticFeatures.score — Method

score(
    f::AcousticComplexityIndex,
    x::MetaArrays.MetaArray{var"#s13",SignalAnalysis.SamplingInfo,T<:Real,N} where N where var"#s13"
) -> Array{_A,1} where _A

Score of x based on Acoustic Complexity Index. jbin is the temporal size of a sub-spectrogram.

Reference:

N. Pieretti, A. Farina, D. Morri, "A new methodology to infer the singing activity of an avian community: The Acoustic Complexity Index (ACI)", 2011.

Examples:

julia> x = Score(AcousticComplexityIndex(1024, 0, 30), randn(960000); fs = 96000)
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1]
    :feature, ["Acoustic Complexity Index"]
    :channel, [1]
And data, a 1×1×1 Array{Float64, 3}:
[:, :, 1] =
 9055.209860502793

 julia> x = Score(AcousticComplexityIndex(1024, 0, 30), randn(960000), winlen=96000, noverlap=48000, fs=96000)
 3-dimensional AxisArray{Float64,3,...} with axes:
     :sample, [1, 48001, 96001, 144001, 192001, 240001, 288001, 336001, 384001, 432001, 480001, 528001, 576001, 624001, 672001, 720001, 768001, 816001, 864001]
     :feature, ["Acoustic Complexity Index"]
     :channel, [1]
 And data, a 19×1×1 Array{Float64, 3}:
 [:, :, 1] =
 874.3303017743206
 ⋮
 875.986826332567

AcousticFeatures.score — Method

score(
    f::AcousticDiversityIndex,
    x::MetaArrays.MetaArray{var"#s13",SignalAnalysis.SamplingInfo,T<:Real,N} where N where var"#s13"
) -> Array{_A,1} where _A

Score of x based on acoustic diversity index.

Reference

Villanueva-Rivera, L. J., B. C. Pijanowski, J. Doucette, and B. Pekin. 2011. A primer of acoustic analysis for landscape ecologists. Landscape Ecology 26: 1233-1246.

https://github.com/patriceguyot/Acoustic_Indices

Examples:

```julia-repl julia> x = Score(AcousticDiversityIndex(256, 128, 50, (50, 1000)), randn(9600); fs = 2000) 3-dimensional AxisArray{Float64,3,...} with axes: :sample, [1] :feature, ["Acoustic Diversity Index"] :channel, [1] And data, a 1×1×1 Array{Float64, 3}: [:, :, 1] = 2.833213240809075

julia> x = Score(AcousticDiversityIndex(256, 128, 50, (50, 1000)), randn(9600); winlen = 960, noverlap = 480, fs = 2000) 3-dimensional AxisArray{Float64,3,...} with axes: :sample, [1, 481, 961, 1441, 1921, 2401, 2881, 3361, 3841, 4321, 4801, 5281, 5761, 6241, 6721, 7201, 7681, 8161, 8641] :feature, ["Acoustic Diversity Index"] :channel, [1] And data, a 19×1×1 Array{Float64, 3}: [:, :, 1] = 2.833213344056216 ⋮ 2.833213344056216

AcousticFeatures.score — Method

score(
    _::Energy,
    x::AbstractArray{T<:Real,1}
) -> Array{_A,1} where _A

Score of x based on mean energy.

Examples:

julia> x = Score(Energy(), randn(9600))
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1]
    :feature, ["Energy"]
    :channel, [1]
And data, a 1×1×1 Array{Float64, 3}:
[:, :, 1] =
 1.001273328811565

julia> x = Score(Energy(), randn(9600); winlen=960, noverlap=480)
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1, 481, 961, 1441, 1921, 2401, 2881, 3361, 3841, 4321, 4801, 5281, 5761, 6241, 6721, 7201, 7681, 8161, 8641]
    :feature, ["Energy"]
    :channel, [1]
And data, a 19×1×1 Array{Float64, 3}:
[:, :, 1] =
 0.9896401818968652
 ⋮
 0.9602549596080265

AcousticFeatures.score — Method

score(
    f::Entropy,
    x::MetaArrays.MetaArray{var"#s13",SignalAnalysis.SamplingInfo,T<:Real,N} where N where var"#s13"
) -> Array{_A,1} where _A

Score of x based on temporal entropy, spectral entropy and entropy index.

Reference:

J. Sueur, A. Farina, A. Gasc, N. Pieretti, S. Pavoine, Acoustic Indices for Biodiversity Assessment and Landscape Investigation, 2014.

Examples:

julia> x = Score(Entropy(96, 48), randn(9600); fs=9600)
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1]
    :feature, ["Temporal Entropy", "Spectral Entropy", "Entropy Index"]
    :channel, [1]
And data, a 1×3×1 Array{Float64, 3}:
[:, :, 1] =
 0.984822  0.998127  0.982977

julia> x = Score(Entropy(96, 48), randn(9600); winlen=960, noverlap=480, fs=9600)
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1, 481, 961, 1441, 1921, 2401, 2881, 3361, 3841, 4321, 4801, 5281, 5761, 6241, 6721, 7201, 7681, 8161, 8641]
    :feature, ["Temporal Entropy", "Spectral Entropy", "Entropy Index"]
    :channel, [1]
And data, a 19×3×1 Array{Float64, 3}:
[:, :, 1] =
 0.980493  0.988433  0.969152
 ⋮                   
 0.979868  0.987179  0.967305

AcousticFeatures.score — Method

score(
    f::FrequencyContours,
    x::MetaArrays.MetaArray{var"#s13",SignalAnalysis.SamplingInfo,T<:Real,N} where N where var"#s13"
) -> Array{_A,1} where _A

Score of x based on frequency contours count.

Reference:

D. Mellinger, R. Morrissey, L. Thomas, J. Yosco, "A method for detecting whistles, moans, and other frequency contour sounds", 2011 J. Acoust. Soc. Am. 129 4055

Examples:

julia> x = Score(FrequencyContours(512, 256, 1.0, 1000.0, 99.0, 1000.0, 0.05), randn(9600); fs=9600)
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1]
    :feature, ["Frequency Contours"]
    :channel, [1]
And data, a 1×1×1 Array{Float64, 3}:
[:, :, 1] =
 0.004539559014267186

julia> x = Score(FrequencyContours(512, 256, 1.0, 1000.0, 99.0, 1000.0, 0.05), randn(9600); winlen=960, noverlap=480, fs=9600)
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1, 481, 961, 1441, 1921, 2401, 2881, 3361, 3841, 4321, 4801, 5281, 5761, 6241, 6721, 7201, 7681, 8161, 8641]
    :feature, ["Frequency Contours"]
    :channel, [1]
And data, a 19×1×1 Array{Float64, 3}:
[:, :, 1] =
 0.0
 ⋮
 0.0

AcousticFeatures.score — Method

score(
    f::ImpulseStats,
    x::MetaArrays.MetaArray{var"#s13",SignalAnalysis.SamplingInfo,T<:Real,N} where N where var"#s13"
) -> Array{_A,1} where _A

Score of x based on number of impulses, mean and variance of inter-impulse intervals. The minimum height of impulses is defined by a+k*b where a is median of the envelope of x and b is median absolute deviation (MAD) of the envelope of x.

Reference:

Matthew W Legg et al., "Analysis of impulsive biological noise due to snapping shrimp as a point process in time", 2007.

Examples:

julia> x = Score(ImpulseStats(10, 0.01), randn(9600); fs = 9600)
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1]
    :feature, ["Nᵢ", "μᵢᵢ", "varᵢᵢ"]
    :channel, [1]
And data, a 1×3×1 Array{Float64, 3}:
[:, :, 1] =
 0.0  NaN  NaN

julia> x = Score(ImpulseStats(10, 0.01), randn(9600); winlen=960, noverlap=480, fs=9600)
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1, 481, 961, 1441, 1921, 2401, 2881, 3361, 3841, 4321, 4801, 5281, 5761, 6241, 6721, 7201, 7681, 8161, 8641]
    :feature, ["Nᵢ", "μᵢᵢ", "varᵢᵢ"]
    :channel, [1]
And data, a 19×3×1 Array{Float64, 3}:
[:, :, 1] =
 0.0  NaN  NaN
 ⋮               
 0.0  NaN  NaN

AcousticFeatures.score — Method

score(
    f::PSD,
    x::MetaArrays.MetaArray{var"#s13",SignalAnalysis.SamplingInfo,T<:Real,N} where N where var"#s13"
) -> Any

Score of x based on power spectral density in dB scale.

Examples:

julia> x = Score(PSD(64, 32, 96000), randn(9600); fs=96000)
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1]
    :feature, ["PSD-0.0Hz", "PSD-1500.0Hz", "PSD-3000.0Hz", "PSD-4500.0Hz", "PSD-6000.0Hz", "PSD-7500.0Hz", "PSD-9000.0Hz", "PSD-10500.0Hz", "PSD-12000.0Hz", "PSD-13500.0Hz"  …  "PSD-34500.0Hz", "PSD-36000.0Hz", "PSD-37500.0Hz", "PSD-39000.0Hz", "PSD-40500.0Hz", "PSD-42000.0Hz", "PSD-43500.0Hz", "PSD-45000.0Hz", "PSD-46500.0Hz", "PSD-48000.0Hz"]
    :channel, [1]
And data, a 1×33×1 Array{Float64, 3}:
[:, :, 1] =
 -50.1861  …  -46.536  -49.8058

julia> x = Score(PSD(64, 32, 96000), randn(9600), winlen=960, noverlap=480, fs=96000)
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1, 481, 961, 1441, 1921, 2401, 2881, 3361, 3841, 4321, 4801, 5281, 5761, 6241, 6721, 7201, 7681, 8161, 8641]
    :feature, ["PSD-0.0Hz", "PSD-1500.0Hz", "PSD-3000.0Hz", "PSD-4500.0Hz", "PSD-6000.0Hz", "PSD-7500.0Hz", "PSD-9000.0Hz", "PSD-10500.0Hz", "PSD-12000.0Hz", "PSD-13500.0Hz"  …  "PSD-34500.0Hz", "PSD-36000.0Hz", "PSD-37500.0Hz", "PSD-39000.0Hz", "PSD-40500.0Hz", "PSD-42000.0Hz", "PSD-43500.0Hz", "PSD-45000.0Hz", "PSD-46500.0Hz", "PSD-48000.0Hz"]
    :channel, [1]
And data, a 19×33×1 Array{Float64, 3}:
[:, :, 1] =
 -52.0638  -48.6761  …  -49.5882
  ⋮                                                                        
 -47.6196  -48.6223     -49.3313

AcousticFeatures.score — Method

score(
    f::PermutationEntropy,
    x::AbstractArray{T<:Real,1}
) -> Array{Float64,1}

Score of x based on permutation entropy.

Reference:

C. Bandt, B. Pompe, "Permutation entropy: a natural complexity measure for time series", Phys. Rev. Lett., 88 (17), 2002

B. Fadlallah, B. Chen, A. Keil, and J. Príncipe, “Weighted-permutation entropy: a complexity measure for time series incorporating amplitude information,” Physical Review E: Statistical, Nonlinear, and Soft Matter Physics, vol. 87, no. 2, Article ID 022911, 2013.

Examples:

julia> x = Score(PermutationEntropy(7), randn(9600))
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1]
    :feature, ["Permutation Entropy"]
    :channel, [1]
And data, a 1×1×1 Array{Float64, 3}:
[:, :, 1] =
 0.964174724342896

julia> x = Score(PermutationEntropy(7), randn(9600), winlen=960, noverlap=480)
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1, 481, 961, 1441, 1921, 2401, 2881, 3361, 3841, 4321, 4801, 5281, 5761, 6241, 6721, 7201, 7681, 8161, 8641]
    :feature, ["Permutation Entropy"]
    :channel, [1]
And data, a 19×1×1 Array{Float64, 3}:
[:, :, 1] =
 0.7892209267573311
 ⋮
 0.7887095679636025

AcousticFeatures.score — Method

score(
    f::SoundPressureLevel,
    x::AbstractArray{T<:Real,1}
) -> Array{_A,1} where _A

Score of x based on Sound Pressure Level (SPL). x is in micropascal. In water, the common reference ref is 1 micropascal. In air, the common reference ref is 20 micropascal.

Examples:

julia> x = Score(SoundPressureLevel(), randn(9600))
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1]
    :feature, ["SPL"]
    :channel, [1]
And data, a 1×1×1 Array{Float64, 3}:
[:, :, 1] =
 -0.023514452445490917

julia> x = Score(SoundPressureLevel(), randn(9600); winlen=960, noverlap=480)
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1, 481, 961, 1441, 1921, 2401, 2881, 3361, 3841, 4321, 4801, 5281, 5761, 6241, 6721, 7201, 7681, 8161, 8641]
    :feature, ["SPL"]
    :channel, [1]
And data, a 19×1×1 Array{Float64, 3}:
[:, :, 1] =
  0.1844558527671365
  ⋮
 -0.12184606185171237

AcousticFeatures.score — Method

score(
    _::SpectralCentroid,
    x::MetaArrays.MetaArray{var"#s13",SignalAnalysis.SamplingInfo,T<:Real,N} where N where var"#s13"
) -> Array{_A,1} where _A

Score of x based on spectral centroid.

Reference:

https://en.wikipedia.org/wiki/Spectral_centroid

Examples:

julia> x = Score(SpectralCentroid(), randn(9600); fs=9600)
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1]
    :feature, ["Spectral Centroid"]
    :channel, [1]
And data, a 1×1×1 Array{Float64, 3}:
[:, :, 1] =
 2406.918897946181

julia> x = Score(SpectralCentroid(), randn(9600); winlen=960, noverlap=480, fs=9600)
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1, 481, 961, 1441, 1921, 2401, 2881, 3361, 3841, 4321, 4801, 5281, 5761, 6241, 6721, 7201, 7681, 8161, 8641]
    :feature, ["Spectral Centroid"]
    :channel, [1]
And data, a 19×1×1 Array{Float64, 3}:
[:, :, 1] =
 2470.7611082156754
 ⋮
 2421.4182047609647

AcousticFeatures.score — Method

score(
    _::SpectralFlatness,
    x::AbstractArray{T<:Real,1}
) -> Array{_A,1} where _A

Score of x based on spectral flatness.

Reference:

https://en.wikipedia.org/wiki/Spectral_flatness

Examples:

julia> x = Score(SpectralFlatness(), randn(9600))
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1]
    :feature, ["Spectral Flatness"]
    :channel, [1]
And data, a 1×1×1 Array{Float64, 3}:
[:, :, 1] =
 0.5703906724982125

julia> x = Score(SpectralFlatness(), randn(9600); winlen=960, noverlap=480)
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1, 481, 961, 1441, 1921, 2401, 2881, 3361, 3841, 4321, 4801, 5281, 5761, 6241, 6721, 7201, 7681, 8161, 8641]
    :feature, ["Spectral Flatness"]
    :channel, [1]
And data, a 19×1×1 Array{Float64, 3}:
[:, :, 1] =
 0.5338443405802898
 ⋮
 0.5704666324591952

AcousticFeatures.score — Method

score(
    f::StatisticalComplexity,
    x::MetaArrays.MetaArray{var"#s13",SignalAnalysis.SamplingInfo,T<:Real,N} where N where var"#s13"
) -> Array{Float64,1}

Score of x based on statistical complexity.

Reference:

Lopez-Ruiz, R., Mancini, H. L., & Calbet, X. (1995). A Statistical Measure of Complexity. Physics Letters A, 209, 321-326.

A. A. B. Pessa, H. V. Ribeiro, ordpy: A Python package for data analysis with permutation entropy and ordinal network methods, Chaos 31, 063110 (2021).

Examples:

julia> x = Score(StatisticalComplexity(7), randn(9600))
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1]
    :feature, ["Statistical Complexity"]
    :channel, [1]
And data, a 1×1×1 Array{Float64, 3}:
[:, :, 1] =
 0.1203920737205638

julia> x = Score(StatisticalComplexity(7), randn(9600); winlen=960, noverlap=480)
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1, 481, 961, 1441, 1921, 2401, 2881, 3361, 3841, 4321, 4801, 5281, 5761, 6241, 6721, 7201, 7681, 8161, 8641]
    :feature, ["Statistical Complexity"]
    :channel, [1]
And data, a 19×1×1 Array{Float64, 3}:
[:, :, 1] =
 0.5155574606285804
 ⋮
 0.515770942278282

AcousticFeatures.score — Method

score(
    _::SymmetricAlphaStableStats,
    x::AbstractArray{T<:Real,1}
) -> Array{_A,1} where _A

Score of x based on the parameters of Symmetric Alpha Stable Distributions. The parameter α measures the impulsiveness while the parameter scale measures the width of the distributions.

Reference:

https://github.com/org-arl/AlphaStableDistributions.jl

Examples:

julia> x = Score(SymmetricAlphaStableStats(), randn(9600))
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1]
    :feature, ["α", "scale"]
    :channel, [1]
And data, a 1×2×1 Array{Float64, 3}:
[:, :, 1] =
 1.97123  0.701582

julia> x = Score(SymmetricAlphaStableStats(), randn(9600); winlen=960, noverlap=480)
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1, 481, 961, 1441, 1921, 2401, 2881, 3361, 3841, 4321, 4801, 5281, 5761, 6241, 6721, 7201, 7681, 8161, 8641]
    :feature, ["α", "scale"]
    :channel, [1]
And data, a 19×2×1 Array{Float64, 3}:
[:, :, 1] =
 1.86815  0.66576
 ⋮        
 2.0      0.678246

AcousticFeatures.score — Method

score(
    _::ZeroCrossingRate,
    x::AbstractArray{T<:Real,1}
) -> Array{_A,1} where _A

Score of x based on zero crossing rate.

Refernce:

https://en.wikipedia.org/wiki/Zero-crossing_rate

Examples:

julia> x = Score(ZeroCrossingRate(), randn(9600))
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1]
    :feature, ["ZCR"]
    :channel, [1]
And data, a 1×1×1 Array{Float64, 3}:
[:, :, 1] =
 0.497239295759975

julia> x = Score(ZeroCrossingRate(), randn(9600); winlen=960, noverlap=480)
3-dimensional AxisArray{Float64,3,...} with axes:
    :sample, [1, 481, 961, 1441, 1921, 2401, 2881, 3361, 3841, 4321, 4801, 5281, 5761, 6241, 6721, 7201, 7681, 8161, 8641]
    :feature, ["ZCR"]
    :channel, [1]
And data, a 19×1×1 Array{Float64, 3}:
[:, :, 1] =
 0.516162669447341
 ⋮
 0.470281543274244