Docstrings

AcousticFeatures.Score — Method

Score(f, 
      x; 
      winlen=length(x), 
      noverlap=0, 
      padtype=:fillzeros, 
      subseqtype=Float64,
      preprocess=identity,
      map=map,
      showprogress=true)

Compute acoustic feature f scores of a time series signal x using sliding windows.

By default, window length winlen is the length of x, i.e., the whole signal is used to compute a score, and overlapping samples noverlap is 0. The padtype specifies the form of padding, and for more information, refer to ImageFiltering.jl. The signal is subject to preliminary processing preprocess. Acoustic feature scores of subseqences can be computed through mapping map. showprogress is used to monitor the computations.

AcousticFeatures.score — Method

Score(f::Myriad{S}, x::AbstractVector{T})

Score of x based on myriad.

Reference:

Mahmood et. al., "Optimal and Near-Optimal Detection in Bursty Impulsive Noise," IEEE Journal of Oceanic Engineering, vol. 42, no. 3, pp. 639–653, 2016.

Examples:

julia> x = Score(Myriad(), randn(9600))
2-dimensional AxisArray{Float64,2,...} with axes:
    :row, 1:1
    :col, ["Myriad"]
And data, a 1×1 Array{Float64,2}:
 27691.956992339285

julia> x = Score(Myriad(), randn(9600); winlen=960, noverlap=480)
2-dimensional AxisArray{Float64,2,...} with axes:
    :row, 1:480:9121
    :col, ["Myriad"]
And data, a 20×1 Array{Float64,2}:
 -5487.396124602646
  1977.7969182956683
  3216.396756712948
     ⋮
  2651.158251224668
  3246.7097026864853

AcousticFeatures.score — Method

score(::Energy, x::AbstractVector{T}) where {T<:Real}

Score of x based on mean energy.

Examples:

julia> x = Score(Energy(), randn(9600))
2-dimensional AxisArray{Float64,2,...} with axes:
    :row, 1:1
    :col, ["Energy"]
And data, a 1×1 Array{Float64,2}:
 0.9960607967861373

julia> x = Score(Energy(), randn(9600); winlen=960, noverlap=480)
2-dimensional AxisArray{Float64,2,...} with axes:
    :row, 1:480:9121
    :col, ["Energy"]
And data, a 20×1 Array{Float64,2}:
 0.5280804987356663
 ⋮
 0.9988797206321275

AcousticFeatures.score — Method

score(f::Entropy, x::AbstractVector{T})

Score of x based on temporal entropy, spectral entropy and entropy index.

Reference:

J. Sueur, A. Farina, A. Gasc, N. Pieretti, S. Pavoine, Acoustic Indices for Biodiversity Assessment and Landscape Investigation, 2014.

Examples:

julia> x = Score(Entropy(9600, 96, 48), randn(9600))
2-dimensional AxisArray{Float64,2,...} with axes:
    :row, 1:1
    :col, ["Temporal Entropy", "Spectral Entropy", "Entropy Index"]
And data, a 1×3 Array{Float64,2}:
 0.984457  0.997608  0.982103

julia> x = Score(Entropy(9600, 96, 48), randn(9600); winlen=960, noverlap=480)
2-dimensional AxisArray{Float64,2,...} with axes:
    :row, 1:480:9121
    :col, ["Temporal Entropy", "Spectral Entropy", "Entropy Index"]
And data, a 20×3 Array{Float64,2}:
 0.903053  0.982299  0.887068
 0.980151  0.986018  0.966446
 0.981492  0.984845  0.966618
 0.980283  0.986635  0.967182
 0.978714  0.987383  0.966366
 ⋮                   
 0.97895   0.986322  0.96556
 0.980274  0.983338  0.963941
 0.980822  0.99296   0.973917
 0.979092  0.989817  0.969122

AcousticFeatures.score — Method

score(f::FrequencyContours, x::AbstractVector{T})

Score of x based on frequency contours count.

Reference:

D. Mellinger, R. Morrissey, L. Thomas, J. Yosco, "A method for detecting whistles, moans, and other frequency contour sounds", 2011 J. Acoust. Soc. Am. 129 4055

Examples:

julia> x = Score(FrequencyContours(9600, 512, 256, 1.0, 1000.0, 99.0, 1000.0, 0.05), randn(9600))
2-dimensional AxisArray{Float64,2,...} with axes:
    :row, 1:1
    :col, ["Frequency Contours"]
And data, a 1×1 Array{Float64,2}:
 0.0038910505836575876

julia> x = Score(FrequencyContours(9600, 512, 256, 1.0, 1000.0, 99.0, 1000.0, 0.05), randn(9600); winlen=960, noverlap=480)
2-dimensional AxisArray{Float64,2,...} with axes:
    :row, 1:480:9121
    :col, ["Frequency Contours"]
And data, a 20×1 Array{Float64,2}:
 0.0
 0.0
 0.0
 0.0
 ⋮
 0.0
 0.0
 0.0

AcousticFeatures.score — Method

score(f::ImpulseStats, x::AbstractVector{T})

Score of x based on number of impulses, mean and variance of inter-impulse intervals. The minimum height of impulses is defined by a+k*b where a is median of the envelope of x and b is median absolute deviation (MAD) of the envelope of x.

Reference:

Matthew W Legg et al., "Analysis of impulsive biological noise due to snapping shrimp as a point process in time", 2007.

Examples:

julia> x = Score(ImpulseStats(9600, 10, 0.01), randn(9600))
2-dimensional AxisArray{Float64,2,...} with axes:
    :row, 1:1
    :col, ["Nᵢ", "μᵢᵢ", "varᵢᵢ"]
And data, a 1×3 Array{Float64,2}:
 0.0  0.0  0.0

julia> x = Score(ImpulseStats(9600, 10, 0.01), randn(9600); winlen=960, noverlap=480)
2-dimensional AxisArray{Float64,2,...} with axes:
    :row, 1:480:9121
    :col, ["Nᵢ", "μᵢᵢ", "varᵢᵢ"]
And data, a 20×3 Array{Float64,2}:
 4.0  0.0140972  0.0621181
 0.0  0.0        0.0
 0.0  0.0        0.0
 0.0  0.0        0.0
 ⋮               
 0.0  0.0        0.0
 0.0  0.0        0.0
 0.0  0.0        0.0

AcousticFeatures.score — Method

score(f::PermutationEntropy, x::AbstractVector{T})

Score of x based on permutation entropy.

Reference:

C. Bandt, B. Pompe, "Permutation entropy: a natural complexity measure for time series", Phys. Rev. Lett., 88 (17), 2002

Examples:

julia> x = Score(PermutationEntropy(7), randn(9600))
2-dimensional AxisArray{Float64,2,...} with axes:
    :row, 1:1
    :col, ["Permutation Entropy"]
And data, a 1×1 Array{Float64,2}:
 0.9637270776723836

julia> x = Score(PermutationEntropy(7), randn(9600), winlen=960, noverlap=480)
2-dimensional AxisArray{Float64,2,...} with axes:
    :row, 1:480:9121
    :col, ["Permutation Entropy"]
And data, a 20×1 Array{Float64,2}:
 0.4432867336969194
 0.7896679491573086
 0.7914368148634888
 0.790455877419772
 ⋮
 0.7894106035823039
 0.7886452315698513
 0.7897322855510599
 0.7884747786386084

AcousticFeatures.score — Method

score(f::SoundPressureLevel, x::AbstractVector{T})

Score of x based on Sound Pressure Level (SPL). x is in micropascal. In water, the common reference ref is 1 micropascal. In air, the common reference ref is 20 micropascal.

Examples:

julia> x = Score(SoundPressureLevel(), randn(9600))
2-dimensional AxisArray{Float64,2,...} with axes:
    :row, 1:1
    :col, ["SPL"]
And data, a 1×1 Array{Float64,2}:
 -0.08307636105819256

julia> x = Score(SoundPressureLevel(), randn(9600); winlen=960, noverlap=480)
2-dimensional AxisArray{Float64,2,...} with axes:
    :row, 1:480:9121
    :col, ["SPL"]
And data, a 20×1 Array{Float64,2}:
 -2.369874304880999
  0.2795371978218069
  ⋮
 -0.04985476533237352
  0.11412307503113574

AcousticFeatures.score — Method

score(f::SpectralCentroid, x::AbstractVector{T})

Score of x based on spectral centroid.

Reference:

https://en.wikipedia.org/wiki/Spectral_centroid

Examples:

julia> x = Score(SpectralCentroid(9600), randn(9600))
2-dimensional AxisArray{Float64,2,...} with axes:
    :row, 1:1
    :col, ["Spectral Centroid"]
And data, a 1×1 Array{Float64,2}:
 2387.4592177121676

julia> x = Score(SpectralCentroid(9600), randn(9600); winlen=960, noverlap=480)
2-dimensional AxisArray{Float64,2,...} with axes:
    :row, 1:480:9121
    :col, ["Spectral Centroid"]
And data, a 20×1 Array{Float64,2}:
 2398.6889311658415
 2362.570125358973
 2342.6919660952803
 2375.5903954079977
    ⋮
 2415.431353476017
 2453.7105902333437
 2449.222535628719
 2380.319011105224

AcousticFeatures.score — Method

score(::SpectralFlatness, x::AbstractVector{T})

Score of x based on spectral flatness.

Reference:

https://en.wikipedia.org/wiki/Spectral_flatness

Examples:

julia> x = Score(SpectralFlatness(), randn(9600))
2-dimensional AxisArray{Float64,2,...} with axes:
    :row, 1:1
    :col, ["Spectral Flatness"]
And data, a 1×1 Array{Float64,2}:
 0.5598932661540399

julia> x = Score(SpectralFlatness(), randn(9600); winlen=960, noverlap=480)
2-dimensional AxisArray{Float64,2,...} with axes:
    :row, 1:480:9121
    :col, ["Spectral Flatness"]
And data, a 20×1 Array{Float64,2}:
 0.5661636483227057
 0.543740942647357
 0.5854629162797854
 0.532148471407988
 ⋮
 0.5451002001200566
 0.5496170417608265
 0.537335859768483
 0.535056705285523

AcousticFeatures.score — Method

score(::SymmetricAlphaStableStats, x::AbstractVector{T})

Score of x based on the parameters of Symmetric Alpha Stable Distributions. The parameter α measures the impulsiveness while the parameter scale measures the width of the distributions.

Reference:

https://github.com/org-arl/AlphaStableDistributions.jl

Examples:

julia> x = Score(SymmetricAlphaStableStats(), randn(9600))
2-dimensional AxisArray{Float64,2,...} with axes:
    :row, 1:1
    :col, ["α", "scale"]
And data, a 1×2 Array{Float64,2}:
 2.0  0.714388

julia> x = Score(SymmetricAlphaStableStats(), randn(9600); winlen=960, noverlap=480)
2-dimensional AxisArray{Float64,2,...} with axes:
    :row, 1:480:9121
    :col, ["α", "scale"]
And data, a 20×2 Array{Float64,2}:
 0.5      0.0
 1.90067  0.663918
 1.83559  0.614218
 ⋮        
 1.80072  0.676852
 1.94506  0.677581

AcousticFeatures.score — Method

score(::ZeroCrossingRate, x::AbstractVector{T})

Score of x based on zero crossing rate.

Refernce:

https://en.wikipedia.org/wiki/Zero-crossing_rate

Examples:

julia> x = Score(ZeroCrossingRate(), randn(9600))
2-dimensional AxisArray{Float64,2,...} with axes:
    :row, 1:1
    :col, ["ZCR"]
And data, a 1×1 Array{Float64,2}:
 0.5027083333333333

julia> x = Score(ZeroCrossingRate(), randn(9600); winlen=960, noverlap=480)
2-dimensional AxisArray{Float64,2,...} with axes:
    :row, 1:480:9121
    :col, ["ZCR"]
And data, a 20×1 Array{Float64,2}:
 0.2375
 0.46979166666666666
 0.4708333333333333
 0.49583333333333335
 0.5072916666666667
 ⋮
 0.5145833333333333
 0.49375
 0.5052083333333334
 0.5125
 0.4947916666666667